ai_service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# Backend & API
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.8.1
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.1.0

# Medical Image I/O
pydicom==3.0.1
# aspose-medical-python-net  # Optional: MIT-licensed & HIPAA-aware

# Image Processing
Pillow==11.2.1
opencv-python-headless==4.11.0.86
scikit-image>=0.25.0

# Scientific Computing
numpy==2.2.4
scipy==1.15.2
scikit-learn==1.6.1

# Visualization & Tracing
matplotlib>=3.10.0
PyQt6>=6.8.0

# Deep Learning
torch==2.6.0+cpu
torchvision==0.21.0+cpu
monai>=1.4.0

# AI & LLM Integration
openai>=1.0.0
google-generativeai>=0.8.3

# Testing
pytest==8.3.5
pytest-asyncio==0.26.0
